#8638 Editors(Lock, Session, View) from ext.exasol to ext.exasol.ui moved, bundles in ext.exasol in ru localized
