+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/META-INF/MANIFEST.MF
@@ -11,7 +11,10 @@ Require-Bundle: org.eclipse.ui,
  org.jkiss.dbeaver.ui,
  org.jkiss.dbeaver.ext.exasol,
  org.jkiss.dbeaver.tasks.ui,
- org.jkiss.dbeaver.ui.editors.sql
+ org.jkiss.dbeaver.ui.editors.sql,
+ org.jkiss.dbeaver.ext.ui.locks,
+ org.jkiss.dbeaver.ui.editors.session,
+ org.jkiss.dbeaver.ui.editors.base
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Vendor: %Bundle-Vendor
