+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle.properties
@@ -8,6 +8,15 @@ Bundle-Vendor = JKISS
 
 dialog.connection.header = Exasol Connection Settings
 
+editor.exasol.source.view.description = Source code
+editor.exasol.source.view.label = Source
+
+editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name = Lock Manager
+editor.org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor.name = Application Manager
+
+editor.session_manager.name = Session Manager
+editor.lock_manager.name = Lock Manager
+
 tool.org.jkiss.dbeaver.ext.exasol.table.export.description = Export Tables, Views or Schemas to CSV
 tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = Export Table(s)
 tool.org.jkiss.dbeaver.ext.exasol.table.import.description = Import Tables or Schemas from CSV
