+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle_de.properties
@@ -2,7 +2,16 @@
 # DBeaver - Universal Database Manager
 # Copyright (C) 2010-2020 DBeaver Corp and others
 # Copyright (C) 2017 Karl Griesser (fullref@gmail.com)
+
 dialog.connection.header = Exasol Verbindungseinstellungen
+
+editor.lock_manager.name    = Sperrmanager
+editor.session_manager.name = Sitzungsmanager
+
+editor.exasol.source.view.description                                        = Quellcode
+editor.exasol.source.view.label                                              = Quelle
+editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name            = Sperrmanager
+editor.org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor.name = Applikationsmanager
 
 tool.org.jkiss.dbeaver.ext.exasol.table.export.description = Exportieren von Tabellen, Ansichten oder Schemas nach CSV
 tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = Tabelle(n) exportieren
