+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle_ja.properties
@@ -5,6 +5,14 @@
 
 dialog.connection.header = \u00A0Exasol\u306E\u63A5\u7D9A\u8A2D\u5B9A
 
+editor.session_manager.name=\u30BB\u30C3\u30B7\u30E7\u30F3
+editor.lock_manager.name=\u30ED\u30C3\u30AF
+
+editor.exasol.source.view.description                                        = \u00A0\u30BD\u30FC\u30B9\u30B3\u30FC\u30C9
+editor.exasol.source.view.label                                              = \u00A0\u30BD\u30FC\u30B9
+editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name            = \u00A0\u30ED\u30C3\u30AF\u30DE\u30CD\u30FC\u30B8\u30E3\u30FC
+editor.org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor.name = \u00A0\u30A2\u30D7\u30EA\u30B1\u30FC\u30B7\u30E7\u30F3\u30DE\u30CD\u30FC\u30B8\u30E3
+
 tool.org.jkiss.dbeaver.ext.exasol.table.export.description = \u00A0\u8868\u3001\u30D3\u30E5\u30FC\u307E\u305F\u306F\u30B9\u30AD\u30FC\u30DE\u3092CSV\u306B\u30A8\u30AF\u30B9\u30DD\u30FC\u30C8\u3059\u308B
 tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = \u00A0\u8F38\u51FA\u8868
 tool.org.jkiss.dbeaver.ext.exasol.table.import.description = \u00A0CSV\u304B\u3089\u30C6\u30FC\u30D6\u30EB\u307E\u305F\u306F\u30B9\u30AD\u30FC\u30DE\u3092\u30A4\u30F3\u30DD\u30FC\u30C8\u3059\u308B
