+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle_pt_BR.properties
@@ -6,7 +6,15 @@
 Bundle-Name = DBeaver Exasol UI
 Bundle-Vendor = JKISS
 
-dialog.connection.header = Configura\u00E7\u00F5es de conex\u00E3o Exasol 
+dialog.connection.header = Configura\u00E7\u00F5es de conex\u00E3o Exasol
+
+editor.session_manager.name = Ger\u00EAnciador de sess\u00E3o
+editor.lock_manager.name = Travamento
+
+editor.exasol.source.view.description                                        = Codigo fonte
+editor.exasol.source.view.label                                              = Fonte
+editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name            = Gerenciador de bloqueio
+editor.org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor.name = Gerenciador de aplica\u00E7\u00E3o
 
 tool.org.jkiss.dbeaver.ext.exasol.table.export.description = Exportar Tabelas, Views ou  Esquema para CSV
 tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = Exportar Tabela(s)
