+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle_ru.properties
@@ -1,6 +1,14 @@
 
 dialog.connection.header = \u041F\u0430\u0440\u0430\u043C\u0435\u0442\u0440\u044B \u043F\u043E\u0434\u043A\u043B\u044E\u0447\u0435\u043D\u0438\u044F Exasol
 
+editor.session_manager.name = \u041C\u0435\u043D\u0435\u0434\u0436\u0435\u0440 \u0441\u0435\u0441\u0441\u0438\u0439
+editor.lock_manager.name = \u0414\u0438\u0441\u043F\u0435\u0442\u0447\u0435\u0440 \u0431\u043B\u043E\u043A\u0438\u0440\u043E\u0432\u043E\u043A
+
+editor.exasol.source.view.description = \u0418\u0441\u0445\u043E\u0434\u043D\u044B\u0439 \u043A\u043E\u0434
+editor.exasol.source.view.label = \u0418\u0441\u0445\u043E\u0434\u043D\u044B\u0439 \u043A\u043E\u0434
+
+editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name = \u0414\u0438\u0441\u043F\u0435\u0442\u0447\u0435\u0440 \u0431\u043B\u043E\u043A\u0438\u0440\u043E\u0432\u043E\u043A
+
 tool.org.jkiss.dbeaver.ext.exasol.table.export.description = \u042D\u043A\u0441\u043F\u043E\u0440\u0442 \u0442\u0430\u0431\u043B\u0438\u0446, \u043F\u0440\u0435\u0434\u0441\u0442\u0430\u0432\u043B\u0435\u043D\u0438\u0439 \u0438\u043B\u0438 \u0441\u0445\u0435\u043C \u0432 CSV
 tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = \u042D\u043A\u0441\u043F\u043E\u0440\u0442 \u0442\u0430\u0431\u043B\u0438\u0446(\u044B)
 tool.org.jkiss.dbeaver.ext.exasol.table.import.description = \u0418\u043C\u043F\u043E\u0440\u0442\u0438\u0440\u043E\u0432\u0430\u0442\u044C \u0442\u0430\u0431\u043B\u0438\u0446\u044B \u0438\u043B\u0438 \u0441\u0445\u0435\u043C\u044B \u0438\u0437 CSV
