+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/OSGI-INF/l10n/bundle_zh.properties
@@ -6,3 +6,11 @@ tool.org.jkiss.dbeaver.ext.exasol.table.export.description = \u5BFC\u51FA\u8868,
 tool.org.jkiss.dbeaver.ext.exasol.table.export.name        = \u5BFC\u51FA\u8868
 tool.org.jkiss.dbeaver.ext.exasol.table.import.description = \u4ECE CSV \u5BFC\u5165\u8868\u6216\u6A21\u5F0F
 tool.org.jkiss.dbeaver.ext.exasol.table.import.name        = \u5BFC\u5165\u8868
+
+editor.session_manager.name = \u4F1A\u8BDD\u7BA1\u7406
+editor.lock_manager.name = \u9501\u53E4\u7BA1\u7406\u5668
+
+editor.exasol.source.view.description                                        = \u6E90\u7801
+editor.exasol.source.view.label                                              = \u6E90
+editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name            = \u9501\u7BA1\u7406\u5668
+editor.org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor.name = \u5E94\u7528\u7BA1\u7406\u5668
