+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/src/org/jkiss/dbeaver/ext/exasol/ui/internal/ExasolResources.properties
@@ -18,6 +18,11 @@ dialog_table_tools_string_sep = String Separator
 dialog_table_tools_string_sep_mode = Use String Separator
 dialog_table_tools_success_title = Success
 
+editors_exasol_session_editor_action_kill = Kill Session
+editors_exasol_session_editor_confirm_action = {0} {1}. Are you sure?
+editors_exasol_session_editor_title_kill_session = Kill Session
+editors_exasol_session_editor_title_kill_session_statement = Terminate Statement
+
 exasol_partition_description = Information about columns that are used for partitioning
 exasol_partition_name = Partition info
