+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/src/org/jkiss/dbeaver/ext/exasol/ui/internal/ExasolResources_de.properties
@@ -23,6 +23,11 @@ dialog_table_tools_string_sep = Zeichentrenneichen
 dialog_table_tools_string_sep_mode = Zeichentrennzeichen Verwendungsart
 dialog_table_tools_success_title = Erfolg
 
+editors_exasol_session_editor_action_kill = Verbindung terminieren
+editors_exasol_session_editor_confirm_action = {0} {1}. Sind Sie sicher?
+editors_exasol_session_editor_title_kill_session = Verbindung terminieren
+editors_exasol_session_editor_title_kill_session_statement = Befehl terminieren
+
 exasol_partition_description = Informationen zu den Partitionierungsspalten
 exasol_partition_name = Partitionsinformation
