+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/src/org/jkiss/dbeaver/ext/exasol/ui/internal/ExasolResources_ja.properties
@@ -17,3 +17,8 @@ dialog_table_tools_row_sep_mode = \u884C\u533A\u5207\u308A\u8A18\u53F7
 dialog_table_tools_string_sep = \u30B9\u30C8\u30EA\u30F3\u30B0\u30BB\u30D1\u30EC\u30FC\u30BF
 dialog_table_tools_string_sep_mode = \u6587\u5B57\u5217\u30BB\u30D1\u30EC\u30FC\u30BF\u3092\u4F7F\u7528\u3059\u308B
 dialog_table_tools_success_title = \u6210\u529F
+
+editors_exasol_session_editor_action_kill = \u00A0\u30BB\u30C3\u30B7\u30E7\u30F3\u3092\u7D42\u4E86
+editors_exasol_session_editor_confirm_action = \u00A0{0} "{1}"\u672C\u5F53\u306B\u3088\u308D\u3057\u3044\u3067\u3059\u304B\uFF1F
+editors_exasol_session_editor_title_kill_session = \u00A0\u30BB\u30C3\u30B7\u30E7\u30F3\u3092\u7D42\u4E86
+editors_exasol_session_editor_title_kill_session_statement = \u00A0Terminate\u30B9\u30C6\u30FC\u30C8\u30E1\u30F3\u30C8
