+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/src/org/jkiss/dbeaver/ext/exasol/ui/internal/ExasolResources_ru.properties
@@ -18,6 +18,10 @@ dialog_table_tools_string_sep_mode = \u0418\u0441\u043F\u043E\u043B\u044C\u0437\
 dialog_table_tools_string_sep = \u0420\u0430\u0437\u0434\u0435\u043B\u0438\u0442\u0435\u043B\u044C \u0441\u0442\u0440\u043E\u043A
 dialog_table_tools_success_title = \u0423\u0441\u043F\u0435\u0448\u043D\u043E
 
+editors_exasol_session_editor_action_kill = \u041F\u0440\u0435\u0440\u0432\u0430\u0442\u044C \u0441\u0435\u0441\u0441\u0438\u044E
+editors_exasol_session_editor_confirm_action = {0} {1}. \u0412\u044B \u0443\u0432\u0435\u0440\u0435\u043D\u044B?
+editors_exasol_session_editor_title_kill_session = \u041F\u0440\u0435\u0440\u0432\u0430\u0442\u044C \u0441\u0435\u0441\u0441\u0438\u044E
+
 label_backup_host_list = \u0421\u043F\u0438\u0441\u043E\u043A \u0440\u0435\u0437\u0435\u0440\u0432\u043D\u044B\u0445 \u0445\u043E\u0441\u0442\u043E\u0432
 label_database = \u0411\u0430\u0437\u0430 \u0434\u0430\u043D\u043D\u044B\u0445
 label_encrypt = \u0428\u0438\u0444\u0440\u043E\u0432\u0430\u043D\u0438\u0435 \u0441\u0432\u044F\u0437\u0438
