+++ plugins/org.jkiss.dbeaver.ext.exasol.ui/src/org/jkiss/dbeaver/ext/exasol/ui/internal/ExasolResources_zh.properties
@@ -20,3 +20,8 @@ dialog_table_tools_row_sep_mode = \u884C\u5206\u9694\u7B26
 dialog_table_tools_string_sep = \u5B57\u7B26\u4E32\u5206\u9694\u7B26
 dialog_table_tools_string_sep_mode = \u4F7F\u7528\u5B57\u7B26\u4E32\u5206\u9694\u7B26
 dialog_table_tools_success_title = \u6210\u529F
+
+editors_exasol_session_editor_action_kill = \u6740\u6389\u4F1A\u8BDD
+editors_exasol_session_editor_confirm_action = {0} "{1}". \u4F60\u786E\u5B9A\u5417?
+editors_exasol_session_editor_title_kill_session = \u6740\u6389\u4F1A\u8BDD
+editors_exasol_session_editor_title_kill_session_statement = \u7EC8\u6B62\u8BED\u53E5
