+++ plugins/org.jkiss.dbeaver.ext.exasol/OSGI-INF/l10n/bundle.properties
@@ -18,10 +18,6 @@ dialog.connection.header = Exasol Connection Settings
 driver.exasol.description = Exasol JDBC driver
 driver.exasol.label       = Exasol
 
-editor.exasol.source.view.description                                        = Source code
-editor.exasol.source.view.label                                              = Source
-editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name            = Lock Manager
-editor.org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor.name = Application Manager
 editor.source.ddl.description                                                = DDL
 editor.source.ddl.name                                                       = DDL
 editor.source.declaration.description                                        = Source
@@ -232,6 +228,11 @@ meta.org.jkiss.dbeaver.ext.exasol.model.security.ExasolUser.passwordState.name  
 meta.org.jkiss.dbeaver.ext.exasol.model.security.ExasolUser.passwordStateChanged.name             = Password status changed
 ##meta.org.jkiss.dbeaver.ext.exasol.model.security.ExasolUser.priority.name                         = Priority
 meta.org.jkiss.dbeaver.ext.exasol.model.security.ExasolViewGrant.view.name                        = View
+meta.org.jkiss.dbeaver.ext.exasol.model.ExasolScript.objectDefinitionText.name = Scripts
+meta.org.jkiss.dbeaver.ext.exasol.model.ExasolView.objectDefinitionText.name = Views
+meta.org.jkiss.dbeaver.ext.exasol.model.ExasolFunction.objectDefinitionText.name = Functions
+meta.org.jkiss.dbeaver.ext.exasol.model.ExasolFunction.sql.name = SQL
+
 
 tool.org.jkiss.dbeaver.ext.exasol.database.reorg.name      = Reorganize Database...
 tool.org.jkiss.dbeaver.ext.exasol.table.export.description = Export Tables, Views or Schemas to CSV
