+++ plugins/org.jkiss.dbeaver.ext.exasol/OSGI-INF/l10n/bundle_de.properties
@@ -14,10 +14,6 @@ dialog.connection.header = Exasol Verbindungseinstellungen
 driver.exasol.description = Exasol JDBC Treiber
 driver.exasol.label       = Exasol
 
-editor.exasol.source.view.description                                        = Quellcode
-editor.exasol.source.view.label                                              = Quelle
-editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name            = Sperrmanager
-editor.org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor.name = Applikationsmanager
 editor.source.ddl.description                                                = DDL
 editor.source.ddl.name                                                       = DDL
 editor.source.declaration.description                                        = Quellcode
