+++ plugins/org.jkiss.dbeaver.ext.exasol/OSGI-INF/l10n/bundle_ja.properties
@@ -14,10 +14,6 @@ dialog.connection.header = \u00A0Exasol\u306E\u63A5\u7D9A\u8A2D\u5B9A
 driver.exasol.description = \u00A0Exasol JDBC\u30C9\u30E9\u30A4\u30D0
 driver.exasol.label       = \u00A0Exasol
 
-editor.exasol.source.view.description                                        = \u00A0\u30BD\u30FC\u30B9\u30B3\u30FC\u30C9
-editor.exasol.source.view.label                                              = \u00A0\u30BD\u30FC\u30B9
-editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name            = \u00A0\u30ED\u30C3\u30AF\u30DE\u30CD\u30FC\u30B8\u30E3\u30FC
-editor.org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor.name = \u00A0\u30A2\u30D7\u30EA\u30B1\u30FC\u30B7\u30E7\u30F3\u30DE\u30CD\u30FC\u30B8\u30E3
 editor.source.ddl.description                                                = \u00A0DDL
 editor.source.ddl.name                                                       = \u00A0DDL
 editor.source.declaration.description                                        = \u00A0\u30BD\u30FC\u30B9
