+++ plugins/org.jkiss.dbeaver.ext.exasol/OSGI-INF/l10n/bundle_pt_BR.properties
@@ -18,10 +18,6 @@ dialog.connection.header = Configura\u00E7\u00F5es de conex\u00E3o Exasol
 driver.exasol.description = Exasol JDBC driver
 driver.exasol.label       = Exasol
 
-editor.exasol.source.view.description                                        = Codigo fonte
-editor.exasol.source.view.label                                              = Fonte
-editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name            = Gerenciador de bloqueio
-editor.org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor.name = Gerenciador de aplica\u00E7\u00E3o
 editor.source.ddl.description                                                = DDL
 editor.source.ddl.name                                                       = DDL
 editor.source.declaration.description                                        = Fonte
