+++ plugins/org.jkiss.dbeaver.ext.exasol/OSGI-INF/l10n/bundle_zh.properties
@@ -12,10 +12,6 @@ dialog.connection.header = Exasol \u8FDE\u63A5\u8BBE\u7F6E
 driver.exasol.description = Exasol JDBC \u9A71\u52A8
 driver.exasol.label       = Exasol
 
-editor.exasol.source.view.description                                        = \u6E90\u7801
-editor.exasol.source.view.label                                              = \u6E90
-editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name            = \u9501\u7BA1\u7406\u5668
-editor.org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor.name = \u5E94\u7528\u7BA1\u7406\u5668
 editor.source.ddl.description                                                = DDL
 editor.source.ddl.name                                                       = DDL
 editor.source.declaration.description                                        = \u6E90
