+++ plugins/org.jkiss.dbeaver.ext.exasol/plugin.xml
@@ -210,14 +210,14 @@
                             label="%tree.sessions.node.name"
                             icon="#admin"
                             description="%tree.sessions.node.description"
-                            editor="org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor"
+                            editor="org.jkiss.dbeaver.ext.exasol.ui.editors.ExasolServerSessionEditor"
                             />
                     <object
                           description="%tree.locks.node.description"
-                          editor="org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor"
+                          editor="org.jkiss.dbeaver.ext.exasol.ui.editors.ExasolLockEditor"
                           icon="#locks"
                           label="%tree.locks.node.name"
-                          type="org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor">
+                          type="org.jkiss.dbeaver.ext.exasol.ui.editors.ExasolLockEditor">
                     </object>
                 </folder>
                 <folder
@@ -490,35 +490,6 @@
               type="folder">
            <objectType name="org.jkiss.dbeaver.ext.exasol.model.ExasolVirtualSchema"/>
         </editor>
-        <editor
-              class="org.jkiss.dbeaver.ext.exasol.editors.ExasolSourceViewEditor"
-              contributor="org.jkiss.dbeaver.ui.editors.sql.SQLEditorContributorNested"
-              description="%editor.exasol.source.view.description"
-              icon="#sql_text"
-              id="exasol.source.view"
-              label="%editor.exasol.source.view.label"
-              position="additions_middle"
-              type="folder">
-           <objectType name="org.jkiss.dbeaver.ext.exasol.model.ExasolView"/>
-           <objectType name="org.jkiss.dbeaver.ext.exasol.model.ExasolScript"/>
-           <objectType name="org.jkiss.dbeaver.ext.exasol.model.ExasolFunction"/>
-        </editor>
-    </extension>
-
-    <extension point="org.eclipse.ui.editors">
-        <editor
-                name="%editor.org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor.name"
-                icon="icons/compile.png"
-                class="org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor"
-                id="org.jkiss.dbeaver.ext.exasol.model.app.ExasolServerSessionEditor"
-                contributorClass="org.jkiss.dbeaver.ui.editors.EditorSearchActionsContributor">
-        </editor>
-        <editor
-            name="%editor.org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor.name"
-            class="org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor"
-            id="org.jkiss.dbeaver.ext.exasol.editors.ExasolLockEditor">
-        </editor>
-
     </extension>
 
     <extension point="org.jkiss.dbeaver.mavenRepository">
