+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/ExasolResources.properties
@@ -47,23 +47,15 @@ dialog_schema_drop_message = Every existing object in the schema will be dropped
 
 dialog_schema_drop_title = Drop Schema Warning
 
-dialog_struct_edit_fk_label_fk_name = Foreign Key Name
-
 edit_exasol_constraint_manager_dialog_title = Create Primary Key
-
-editors_exasol_session_editor_action_kill = Kill Session
-
-editors_exasol_session_editor_confirm_action = {0} {1}. Are you sure?
-
-editors_exasol_session_editor_title_kill_session = Kill Session
-
-editors_exasol_session_editor_title_kill_session_statement = Terminate Statement
 
 exasol_security_policy_description = System wide password security policy
 
 exasol_security_policy_name = System Security Policy
 
 manager_assign_priority_group = Assign priority group
+
+manager_consumer_alter = Alter consumer group
 
 manager_consumer_create = Create new consumer group
