+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/ExasolResources_de.properties
@@ -41,23 +41,15 @@ dialog_priority_group_name = Priorit\u00E4tsgruppenname
 
 dialog_priority_group_weight = Gewicht der Priorit\u00E4tsgruppe
 
-dialog_struct_edit_fk_label_fk_name = Fremschl\u00FCsselname
-
 edit_exasol_constraint_manager_dialog_title = Prim\u00E4rschl\u00FCssel Anlegen
-
-editors_exasol_session_editor_action_kill = Verbindung terminieren
-
-editors_exasol_session_editor_confirm_action = {0} {1}. Sind Sie sicher?
-
-editors_exasol_session_editor_title_kill_session = Verbindung terminieren
-
-editors_exasol_session_editor_title_kill_session_statement = Befehl terminieren
 
 exasol_security_policy_description = Systemweite Passwortsicherheitsrichtlinie
 
 exasol_security_policy_name = Systemsicherheitsrichtlinie
 
 manager_assign_priority_group = Priorit\u00E4tsgruppe zuweisen
+
+manager_consumer_alter = \u00C4ndern Consumer Gruppe
 
 manager_consumer_create = Neue Consumer Gruppe anlegen
