+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/ExasolResources_ja.properties
@@ -13,12 +13,4 @@ dialog_struct_edit_fk_label_fk_name = \u5916\u90E8\u30AD\u30FC\u540D
 
 edit_exasol_constraint_manager_dialog_title = \u4E3B\u30AD\u30FC\u306E\u4F5C\u6210
 
-editors_exasol_session_editor_action_kill = \u00A0\u30BB\u30C3\u30B7\u30E7\u30F3\u3092\u7D42\u4E86
-
-editors_exasol_session_editor_confirm_action = \u00A0{0} "{1}"\u672C\u5F53\u306B\u3088\u308D\u3057\u3044\u3067\u3059\u304B\uFF1F
-
-editors_exasol_session_editor_title_kill_session = \u00A0\u30BB\u30C3\u30B7\u30E7\u30F3\u3092\u7D42\u4E86
-
-editors_exasol_session_editor_title_kill_session_statement = \u00A0Terminate\u30B9\u30C6\u30FC\u30C8\u30E1\u30F3\u30C8
-
 read_table_details = \u30C6\u30FC\u30D6\u30EB\u306E\u8A73\u7D30\u3092\u8AAD\u3080
