+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/ExasolResources_ru.properties
@@ -1,1 +1,3 @@
 exasol_security_policy_name = \u041F\u043E\u043B\u0438\u0442\u0438\u043A\u0430 \u0431\u0435\u0437\u043E\u043F\u0430\u0441\u043D\u043E\u0441\u0442\u0438 \u0441\u0438\u0441\u0442\u0435\u043C\u044B
+
+manager_consumer_alter = \u0418\u0437\u043C\u0435\u043D\u0438\u0442\u044C \u043F\u043E\u0442\u0440\u0435\u0431\u0438\u0442\u0435\u043B\u044C\u0441\u043A\u0443\u044E \u0433\u0440\u0443\u043F\u043F\u0443
