+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/ExasolResources_zh.properties
@@ -14,11 +14,3 @@ dialog_schema_drop_title = Schema\u5220\u9664\u8B66\u544A
 dialog_struct_edit_fk_label_fk_name = \u5916\u952E\u540D\u79F0
 
 edit_exasol_constraint_manager_dialog_title = \u521B\u5EFA\u4E3B\u952E
-
-editors_exasol_session_editor_action_kill = \u6740\u6389\u4F1A\u8BDD
-
-editors_exasol_session_editor_confirm_action = {0} "{1}". \u4F60\u786E\u5B9A\u5417?
-
-editors_exasol_session_editor_title_kill_session = \u6740\u6389\u4F1A\u8BDD
-
-editors_exasol_session_editor_title_kill_session_statement = \u7EC8\u6B62\u8BED\u53E5
